Remove archassault.
I went on IRC to ask what is happening and a guy named arch3y told me there is
some drama. They revoked their gpg master key and the website is down.
In other words: it's dead, the archassault project is dead.

As a future alternative, this guy told me to wait for ArchStrike to come.

Thanks arch3y for answering all my questions and good luck for ArchStrike!